+++ ErsatzTV.Core/Scheduling/PlayoutParameters.cs
@@ -1,0 +1,8 @@
+﻿using ErsatzTV.Core.Domain;
+
+namespace ErsatzTV.Core.Scheduling;
+
+public record PlayoutParameters(
+    DateTimeOffset Start,
+    DateTimeOffset Finish,
+    Map<CollectionKey, List<MediaItem>> CollectionMediaItems);